Better handling of configuration errors
* Use "log" package instead of hard-coded log destination.
  This allows library users to integrate error messages from
  this library with their own log destination or other log libraries
  like github.com/golang/glog.
* Avoid directly terminating the current process with os.Exit.
  Use panic instead.
  In the original design with os.Exit, library users did not have any
  opportunity to cleanup resources (e.g. temporary files, HTTP session
  in other goroutines, ...) nor propagate the error message to the
  parent process.

NOTE: As explained in Effective Go (https://golang.org/doc/effective_go.html#panic),
usually it is better to avoid panic in libraries. But it is not covered
in this commit because it requires incompatible signature change of
Config, ConfigWithEnv and ManualConfig and forces library users to fix
their applications.